requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
eventlet!=0.18.3,>=0.18.2 # MIT
alembic>=0.8.4 # MIT
Babel>=2.3.4 # BSD
docker-py<1.8.0,>=1.6.0 # Apache-2.0
httplib2>=0.7.5 # MIT
iso8601>=0.1.11 # MIT
kombu>=3.0.25 # BSD
netaddr!=0.7.16,>=0.7.12 # BSD
oslo.config>=3.14.0 # Apache-2.0
oslo.context>=2.9.0 # Apache-2.0
oslo.db>=4.10.0 # Apache-2.0
oslo.messaging>=5.2.0 # Apache-2.0
oslo.middleware>=3.0.0 # Apache-2.0
oslo.serialization>=1.10.0 # Apache-2.0
oslo.utils>=3.16.0 # Apache-2.0
oslo.log>=1.14.0 # Apache-2.0
pbr>=1.6 # Apache-2.0
pecan!=1.0.2,!=1.0.3,!=1.0.4,>=1.0.0 # BSD
pycrypto>=2.6 # Public Domain
python-barbicanclient>=4.0.0 # Apache-2.0
python-glanceclient!=2.4.0,>=2.3.0 # Apache-2.0
python-heatclient>=1.4.0 # Apache-2.0
python-zaqarclient>=1.0.0 # Apache-2.0
python-neutronclient>=5.1.0 # Apache-2.0
python-keystoneclient!=2.1.0,>=2.0.0 # Apache-2.0
python-swiftclient>=2.2.0 # Apache-2.0
six>=1.9.0 # MIT
SQLAlchemy<1.1.0,>=1.0.10 # MIT
WSME>=0.8 # MIT
python-mistralclient>=2.0.0 # Apache-2.0
keystonemiddleware!=4.1.0,!=4.5.0,>=4.0.0 # Apache-2.0
